backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.10.1
deepgram-sdk==2.12.0
google-generativeai==0.3.2
elevenlabs==0.2.26
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
cors==1.0.1
fastapi-cors==0.0.6 